requirements.txt
aiohttp>=2.3.6
aiosmtplib>=1.0.2
PyYAML>=3.12